Fix bench coverage summary checks
