+++ src/commands/bench_summary.rs
@@ -12,7 +12,7 @@
 //! to the `--output` file unchanged, so no data is lost — only the default
 //! human presentation becomes compact.
 //!
-//! Artifact pointers (shared-state files, WP Codebox bundle paths,
+//! Artifact pointers (shared-state files, runner bundle paths,
 //! scenario-specific artifacts) and the `homeboy runs show` follow-up
 //! command are surfaced near the top so they are easy to find (#3260).
 
@@ -21,6 +21,9 @@ use std::collections::{BTreeMap, HashMap};
 use serde_json::Value;
 
 use super::summary_json::{array_len, string_value, u64_value, usize_value, value_at};
+
+mod coverage;
+pub(crate) use self::coverage::bench_coverage_lines;
 
 /// Render a compact summary for a serialized `BenchOutput` value. Returns
 /// `None` for variants where the full JSON is the better default (lists,
@@ -244,228 +247,6 @@ pub(crate) fn bench_hotspot_lines(output: &Value) -> Vec<String> {
     lines
 }
 
-/// Render generic bench/fuzzer coverage metadata when runners provide it.
-///
-/// The extractor is intentionally schema-blind. It does not know product
-/// domains or artifact schemas; it only looks for generic coverage keys and
-/// count fields commonly present in bench result metadata.
-pub(crate) fn bench_coverage_lines(output: &Value) -> Vec<String> {
-    let Some(summary) = collect_coverage_summary(output) else {
-        return Vec::new();
-    };
-
-    let mut lines = vec!["Coverage:".to_string()];
-    let mut counts = Vec::new();
-    if let Some(value) = summary.surface_count {
-        counts.push(format!("discovered={value}"));
-    }
-    if let Some(value) = summary.exercised_count {
-        counts.push(format!("exercised={value}"));
-    }
-    if let Some(value) = summary.skipped_count {
-        counts.push(format!("skipped_unsafe={value}"));
-    }
-    if let Some(value) = summary.failed_count {
-        counts.push(format!("failed={value}"));
-    }
-    if !counts.is_empty() {
-        lines.push(format!("  Surfaces: {}", counts.join(" ")));
-    }
-    if let Some(value) = summary.coverage_gap_count {
-        lines.push(format!("  Coverage gaps: {value}"));
-    }
-    if !summary.top_uncovered_groups.is_empty() {
-        lines.push("  Top uncovered groups:".to_string());
-        for group in summary.top_uncovered_groups {
-            match group.count {
-                Some(count) => lines.push(format!("    {}: {count}", group.name)),
-                None => lines.push(format!("    {}", group.name)),
-            }
-        }
-    }
-
-    if lines.len() == 1 {
-        return Vec::new();
-    }
-    lines
-}
-
-#[derive(Debug, Default)]
-struct CoverageSummary {
-    surface_count: Option<usize>,
-    exercised_count: Option<usize>,
-    skipped_count: Option<usize>,
-    failed_count: Option<usize>,
-    coverage_gap_count: Option<usize>,
-    top_uncovered_groups: Vec<UncoveredGroup>,
-}
-
-#[derive(Debug)]
-struct UncoveredGroup {
-    name: String,
-    count: Option<usize>,
-}
-
-fn collect_coverage_summary(output: &Value) -> Option<CoverageSummary> {
-    let mut summary = CoverageSummary::default();
-    let candidates = coverage_candidates(output);
-    if candidates.is_empty() {
-        return None;
-    }
-
-    for candidate in &candidates {
-        if let Some(value) = usize_value(candidate, &["surface_count"]) {
-            summary.surface_count.get_or_insert(value);
-        }
-        if let Some(value) = usize_value(candidate, &["exercised_count"]) {
-            summary.exercised_count.get_or_insert(value);
-        }
-        if let Some(value) = usize_value(candidate, &["skipped_count"]) {
-            summary.skipped_count.get_or_insert(value);
-        }
-        if let Some(value) = usize_value(candidate, &["failed_count"]) {
-            summary.failed_count.get_or_insert(value);
-        }
-        if summary.coverage_gap_count.is_none() {
-            summary.coverage_gap_count = array_len(candidate, &["coverage_gaps"])
-                .or_else(|| usize_value(candidate, &["coverage_gap_count"]));
-        }
-        if summary.top_uncovered_groups.is_empty() {
-            summary.top_uncovered_groups = top_uncovered_groups(candidate);
-        }
-    }
-
-    if summary.surface_count.is_none()
-        && summary.exercised_count.is_none()
-        && summary.skipped_count.is_none()
-        && summary.failed_count.is_none()
-        && summary.coverage_gap_count.is_none()
-        && summary.top_uncovered_groups.is_empty()
-    {
-        return None;
-    }
-    Some(summary)
-}
-
-fn coverage_candidates(output: &Value) -> Vec<&Value> {
-    let mut candidates = Vec::new();
-    for path in [
-        vec!["coverage_summary"],
-        vec!["results", "coverage_summary"],
-        vec!["run_metadata", "coverage_summary"],
-        vec!["results", "run_metadata", "coverage_summary"],
-        vec!["metadata", "coverage_summary"],
-    ] {
-        if let Some(value) = value_at(output, &path) {
-            candidates.push(value);
-        }
-    }
-    for path in [
-        Vec::<&str>::new(),
-        vec!["results"],
-        vec!["run_metadata"],
-        vec!["results", "run_metadata"],
-        vec!["metadata"],
-    ] {
-        if let Some(value) = value_at(output, &path) {
-            candidates.push(value);
-        } else if path.is_empty() {
-            candidates.push(output);
-        }
-    }
-    for path in [
-        vec!["artifacts"],
-        vec!["results", "artifacts"],
-        vec!["metadata", "artifacts"],
-    ] {
-        for artifact in value_at(output, &path)
-            .and_then(Value::as_array)
-            .into_iter()
-            .flatten()
-        {
-            if let Some(value) = value_at(artifact, &["coverage_summary"]) {
-                candidates.push(value);
-            }
-            candidates.push(artifact);
-        }
-    }
-    candidates
-}
-
-fn top_uncovered_groups(candidate: &Value) -> Vec<UncoveredGroup> {
-    for key in ["top_uncovered_groups", "uncovered_groups"] {
-        if let Some(groups) = value_at(candidate, &[key]).and_then(Value::as_array) {
-            let groups = groups
-                .iter()
-                .filter_map(group_from_value)
-                .take(5)
-                .collect::<Vec<_>>();
-            if !groups.is_empty() {
-                return groups;
-            }
-        }
-    }
-
-    let Some(gaps) = value_at(candidate, &["coverage_gaps"]).and_then(Value::as_array) else {
-        return Vec::new();
-    };
-    let mut counts: BTreeMap<String, usize> = BTreeMap::new();
-    for gap in gaps {
-        if let Some(group) = gap_group(gap) {
-            *counts.entry(group).or_default() += 1;
-        }
-    }
-    let mut groups = counts
-        .into_iter()
-        .map(|(name, count)| UncoveredGroup {
-            name,
-            count: Some(count),
-        })
-        .collect::<Vec<_>>();
-    groups.sort_by(|a, b| {
-        b.count
-            .unwrap_or(0)
-            .cmp(&a.count.unwrap_or(0))
-            .then_with(|| a.name.cmp(&b.name))
-    });
-    groups.truncate(5);
-    groups
-}
-
-fn group_from_value(value: &Value) -> Option<UncoveredGroup> {
-    if let Some(name) = value.as_str() {
-        return Some(UncoveredGroup {
-            name: name.to_string(),
-            count: None,
-        });
-    }
-    let name = string_value(value, &["group"])
-        .or_else(|| string_value(value, &["name"]))
-        .or_else(|| string_value(value, &["id"]))?;
-    Some(UncoveredGroup {
-        name: name.to_string(),
-        count: usize_value(value, &["count"]).or_else(|| usize_value(value, &["uncovered_count"])),
-    })
-}
-
-fn gap_group(value: &Value) -> Option<String> {
-    if let Some(group) = string_value(value, &["group"])
-        .or_else(|| string_value(value, &["surface_group"]))
-        .or_else(|| string_value(value, &["category"]))
-    {
-        return Some(group.to_string());
-    }
-    let text = value.as_str()?;
-    for separator in ["::", ":", "/", "."] {
-        if let Some((group, _)) = text.split_once(separator) {
-            if !group.is_empty() {
-                return Some(group.to_string());
-            }
-        }
-    }
-    Some(text.to_string())
-}
-
 fn collect_bench_metric_points(output: &Value) -> Vec<(String, String, f64)> {
     let scenarios = value_at(output, &["results", "scenarios"])
         .and_then(Value::as_array)
@@ -623,7 +404,7 @@ fn gate_failure_lines(output: &Value) -> Vec<String> {
     lines
 }
 
-/// Surface artifact pointers (#3260): shared-state files, WP Codebox bundle
+/// Surface artifact pointers (#3260): shared-state files, runner bundle
 /// paths, and scenario-specific artifacts. Each ref carries a path or a URL;
 /// we print whichever locates it, prefixed by scenario + name so it is
 /// grep-friendly.
@@ -822,7 +603,7 @@ mod tests {
                 {
                     "scenario_id": "rtc-smoke",
                     "name": "bundle",
-                    "url": "https://codebox.test/bundle.zip"
+                    "url": "https://runner.test/bundle.zip"
                 }
             ],
             "persisted_run": {
@@ -839,7 +620,7 @@ mod tests {
                     "iterations": 10,
                     "runs": 1,
                     "concurrency": 1,
-                    "runner": { "extension": "wp-codebox", "path": "/runner" }
+                    "runner": { "extension": "sample-runner", "path": "/runner" }
                 },
                 "scenarios": [
                     {
@@ -857,7 +638,7 @@ mod tests {
         assert!(summary.contains("Component: homeboy\n"));
         assert!(summary.contains("Iterations: 10\n"));
         assert!(summary.contains("Run: bench-run-42\n"));
-        assert!(summary.contains("Runner: wp-codebox\n"));
+        assert!(summary.contains("Runner: sample-runner\n"));
         assert!(summary.contains("Scenarios (1): rtc-smoke\n"));
         assert!(summary.contains("Key metrics:\n"));
         assert!(summary.contains("p50_ms=12.5"));
@@ -868,7 +649,7 @@ mod tests {
         // Surface artifact pointers (#3260).
         assert!(summary.contains("Artifacts (2):\n"));
         assert!(summary.contains("rtc-smoke/response-rows: /tmp/shared/response-rows.json\n"));
-        assert!(summary.contains("rtc-smoke/bundle: https://codebox.test/bundle.zip\n"));
+        assert!(summary.contains("rtc-smoke/bundle: https://runner.test/bundle.zip\n"));
         // Point at the persisted run for full evidence.
         assert!(summary.contains("Inspect: homeboy runs show bench-run-42\n"));
         assert!(summary.contains("Artifacts: homeboy runs artifacts bench-run-42\n"));
@@ -970,9 +751,8 @@ mod tests {
         let summary = render_bench_summary(&payload).expect("summary");
 
         assert!(summary.contains("Coverage:\n"));
-        assert!(summary.contains(
-            "  Surfaces: discovered=120 exercised=82 skipped_unsafe=14 failed=3\n"
-        ));
+        assert!(summary
+            .contains("  Surfaces: discovered=120 exercised=82 skipped_unsafe=14 failed=3\n"));
         assert!(summary.contains("  Coverage gaps: 3\n"));
         assert!(summary.contains("  Top uncovered groups:\n"));
         assert!(summary.contains("    runner: 2\n"));
